backend/pyproject.toml
[project]
name = "agentic-rag-backend"
version = "0.1.0"
description = "Backend service for Agentic RAG."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "pydantic-settings>=2.6.0,<3.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "sqlalchemy[asyncio]>=2.0.0,<3.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "pgvector>=0.3.0,<1.0.0",
    "alembic>=1.13.0,<2.0.0",
    "openai>=1.50.0,<2.0.0",
    "tiktoken>=0.7.0,<1.0.0",
    "sse-starlette>=2.0.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0,<1.0.0",
    "pytest>=8.3.0,<9.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
    "ruff>=0.8.0,<1.0.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["B008"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
